permissions

+++ routes/api.php
@@ -26,17 +26,16 @@ Route::group(['prefix' => 'v1'], function () {
     //User Controller
     Route::controller(UserController::class)->group(function () {
         Route::post('login', 'login');
-        Route::post('register', 'register')->name('register');
         //auth required
         Route::group(['middleware' => ['auth:sanctum']], function () {
             Route::get('user', 'getUser');
             //admin
             Route::group(['middleware' => ['role:admin']], function () {
-                Route::get('users', 'users');
-                Route::get('set-user-active/{user}', 'setUserActive');
-                Route::get('set-user-block/{user}', 'setUserBlock');
-                Route::put('users/{user}', 'update');
-                Route::delete('users/{user}', 'destroy');
+                //create officer
+                Route::post('register', 'register')->name('register')->middleware(['check.permission:اضافة الضباط']);
+                Route::get('users', 'users')->middleware(['check.permission:عرض الضباط']);
+                Route::put('users/{user}', 'update')->middleware(['check.permission:تعديل الضباط']);
+                Route::delete('users/{user}', 'destroy')->middleware(['check.permission:حذف الضباط']);
                 Route::get('permissions', 'permissions');
             });
             //all
@@ -47,23 +46,31 @@ Route::group(['prefix' => 'v1'], function () {
     Route::group(['middleware' => ['auth:sanctum']], function () {
         //admin
         Route::group(['middleware' => ['role:admin']], function () {
-            Route::post('set-unit-officer', [UnitController::class, 'setUnitOfficer']);
+            Route::post('set-unit-officer', [UnitController::class, 'setUnitOfficer'])->middleware(['check.permission:تعديل الوحدات']);
+            //units crud
             Route::controller(UnitController::class)->group(function () {
-                Route::get('units/', 'index');
-                Route::post('units/', 'store');
-                Route::put('units/{unit}', 'update');
-                Route::delete('units/{unit}', 'destroy');
+                Route::get('units/', 'index')->middleware(['check.permission:عرض الوحدات']);
+                Route::post('units/', 'store')->middleware(['check.permission:اضافة الوحدات']);
+                Route::put('units/{unit}', 'update')->middleware(['check.permission:تعديل الوحدات']);
+                Route::delete('units/{unit}', 'destroy')->middleware(['check.permission:حذف الوحدات الوحدات']);
             });
-            Route::resource('violations', ViolationController::class)->except(['show', 'edit', 'index']);
-            Route::get('unit-violations', [UnitViolationController::class, 'index']);
+            //violations crud
+            Route::controller(ViolationController::class)->group(function () {
+                Route::get('violations', 'index')->middleware(['check.permission:عرض عناوين المخالفات']);
+                Route::post('violations', 'store')->middleware(['check.permission:اضافة عناوين المخالفات']);
+                Route::put('violations/{violation}', 'update')->middleware(['check.permission:تعديل عناوين المخالفات']);
+                Route::delete('violations/{violation}', 'destroy')->middleware(['check.permission:حذف عناوين المخالفات']);
+            });
+            // Total unit violations
+            Route::get('unit-violations', [UnitViolationController::class, 'index'])->middleware(['check.permission:عرض اجمالي المخالفات']);
         });
-        Route::get('violations', [ViolationController::class, 'index']);
         //all
         Route::middleware(['have.unit'])->group(function () {
-            Route::get('units/{unit}', [UnitController::class, 'show'])->middleware(['can:عرض المخالفات']);
-            Route::post('unit-violations', [UnitViolationController::class, 'store'])->middleware(['can:تسجيل المخالفات']);
-            Route::put('unit-violations/{id}', [UnitViolationController::class, 'update'])->middleware(['can:تعديل المخالفات']);
-            Route::delete('unit-violations/{id}', [UnitViolationController::class, 'destroy'])->middleware(['can:حذف المخالفات']);
+            // One unit violations
+            Route::get('units/{unit}', [UnitController::class, 'show'])->middleware(['check.permission:عرض مخالفات الوحدات']);
+            Route::post('unit-violations', [UnitViolationController::class, 'store'])->middleware(['check.permission:اضافة مخالفات الوحدات']);
+            Route::put('unit-violations/{id}', [UnitViolationController::class, 'update'])->middleware(['check.permission:تعديل مخالفات الوحدات']);
+            Route::delete('unit-violations/{id}', [UnitViolationController::class, 'destroy'])->middleware(['check.permission:حذف مخالفات الوحدات']);
         });
     });
 });
